Support Sidekiq v7
The main change affecting honeykiq is the change of the underlying redis client.

Sidekiq v7 also drops support for Ruby 2.6. We also drop it in this change.

We also bump the example versions and add a reporter which allow us to
test the redis client change in both Sidekiq versions.